backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-backend"
version = "1.0.0"
description = "RAG backend for Physical AI & Humanoid Robotics book"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "M. Owais Abdullah", email = "mrowaisabdullah@gmail.com"},
]
keywords = ["rag", "retrieval", "openai", "qdrant", "fastapi", "robotics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core Dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.2.1",
    "itsdangerous>=2.1.0",
    # OpenAI Integration
    "openai>=1.68.0",
    "openai-agents>=0.2.9",
    "tiktoken>=0.5.2",
    # Vector Database
    "qdrant-client>=1.7.0",
    # Async HTTP Client
    "httpx>=0.26.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.2.1",
    # Rate Limiting
    "slowapi>=0.1.9",
    "limits>=3.13.1",
    # CORS and Security
    "python-multipart>=0.0.6",
    # Environment Configuration
    "python-dotenv>=1.0.0",
    # Logging and Monitoring
    "structlog>=23.2.0",
    "backoff>=2.2.1",
    "python-json-logger>=2.0.7",
    "PyYAML>=6.0.1",
    # Monitoring and Performance
    "psutil>=5.9.6",
    "openai-chatkit>=1.4.0",
    "email-validator>=2.3.0",
    "bcrypt==4.2.0",
    "google-genai>=0.3.0",
    "redis>=7.1.0",
    "python-json-logger>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "aiosmtplib>=5.0.0",
]

[project.optional-dependencies]
dev = [
    # Development Dependencies
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "mypy>=1.8.0",
    "ruff>=0.1.8",
    "pre-commit>=3.5.0",

    # Documentation
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.8",

    # Optional Performance Boost (Unix only)
    "uvloop>=0.19.0; sys_platform != 'win32'",
]

test = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "faker>=20.1.0",
]

[project.urls]
Homepage = "https://github.com/mrowaisabdullah/ai-humanoid-robotics"
Documentation = "https://mrowaisabdullah.github.io/ai-humanoid-robotics/"
Repository = "https://github.com/mrowaisabdullah/ai-humanoid-robotics.git"
"Bug Tracker" = "https://github.com/mrowaisabdullah/ai-humanoid-robotics/issues"

[project.scripts]
rag-server = "rag.main:main"
rag-ingest = "rag.scripts.ingest:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "mypy>=1.8.0",
    "ruff>=0.1.8",
    "pre-commit>=3.5.0",
]

[tool.uv.sources]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "slowapi.*",
    "limits.*",
    "tiktoken.*",
    "structlog.*",
    "psutil.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["rag"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Hatch build configuration
[tool.hatch.build.targets.wheel]
packages = ["rag"]
exclude = [
    "tests",
    "scripts",
    "book_content",
    "logs"
]

[tool.hatch.build.targets.sdist]
include = [
    "/rag",
    "/pyproject.toml",
    "/README.md",
    "/LICENSE"
]
exclude = [
    "tests",
    "scripts",
    "book_content",
    "logs",
    ".git*",
    "*.md",
    "!README.md"
]
